Add Support to frameAncestors CSP directive in order to allow the iframing of gateway in Support-Frontend (#3270)
<!-- See
https://github.com/guardian/recommendations/blob/main/pull-requests.md
for recommendations on raising and reviewing pull requests. -->

## What does this change?
This permits the Iframing of Gateway in Support-Frontend (LOCAL, CODE,
PROD) by passing the appropriate url to frameAncestors CSP Directive.

This PR exists as part of the support onboarding journey we want to
encourage users to register or sign in after they have made a purchase.

To do this we want to iframe in the signin journey into the thankyou
page.

We need to add profile.theguardian.com (more specifically the correct
profile domains for the particular stage that support and profile are
running on) into the frameAncestors CSP directive to allow iframing from
the correct domain.



<!-- A PR should have enough detail to be understandable far in the
future. e.g what is the problem/why is the change needed, how does it
solve it and any questions or points of discussion. Prefer copying
information from a Trello card over linking to it; the card may not
always exist and reviewers may not have access to the board. -->

## How to test

This has been tested locally using Support-Frontend.

<!-- Provide instructions to help others verify the change. This could
take the form of "On PROD, do X and witness Y. On this branch, do X and
witness Z. " -->

## Screenshots
<img width="692" height="560" alt="Screenshot 2025-09-23 at 15 44 53"
src="https://github.com/user-attachments/assets/0903e668-3876-42bc-b6e1-4dcd48f95808"
/>

Co-authored-by: Richard Bangay <richard.bangay@gmail.com>