Make the /signin route a combined signin or register journey (#3260)
## What does this change?
This pr adds the ability to register as well as sign in when using the
`profile.theguardian.com/signin` url. This should hopefully help out
where users that thought they had an account but didn't try to sign in
via the /signin page. Previously these users would not have been sent an
email which can be confusing.

## Ticket

https://trello.com/c/Ar6bvvIN/1673-gateway-combine-sign-in-and-create-account-flows

## Have we considered potential risks?
Without mitigation there is the possibility of introducing user
enumeration into the combined sign in / register flow. All mentions of
whether the user is trying to sign in or register should not be
displayed to the client until they have been authenticated eg after the
passcode page.

## Images
(before and after journey of a user without an account that goes to
`profile.theguardian.com/signin`)
Before             |  After
:-------------------------:|:-------------------------:

![](https://github.com/user-attachments/assets/715a91ff-c8fc-45b8-b656-ee84374fee58)
|
![](https://github.com/user-attachments/assets/5ef6e368-2581-49e6-8dd6-f7651d73b19e)

Before             |  After
:-------------------------:|:-------------------------:

![](https://github.com/user-attachments/assets/cbb0078f-171b-404d-85f5-fbe45cbbb302)
|
![](https://github.com/user-attachments/assets/8bce1b30-370f-4aeb-aad6-3bca5cbd9220)

Before             |  After
:-------------------------:|:-------------------------:

![](https://github.com/user-attachments/assets/a513ae67-cb52-46cc-acfa-a3a45eed3972)
|
![](https://github.com/user-attachments/assets/68146b03-3ced-41f1-99ad-94fd7ee54bce)

Before             |  After
:-------------------------:|:-------------------------:

![](https://github.com/user-attachments/assets/2c2590bb-0173-412a-b92a-7e43f77120dd)
|
![](https://github.com/user-attachments/assets/0b329f3d-a455-49ed-a2aa-89fa4003ba62)

Before             |  After
:-------------------------:|:-------------------------:

![](https://github.com/user-attachments/assets/2c2590bb-0173-412a-b92a-7e43f77120dd)
|
![](https://github.com/user-attachments/assets/81311b5a-a37b-49d1-b9f1-0dc946aefb59)

---------

Co-authored-by: Richard Bangay <richard.bangay@guardian.co.uk>