combine sign in and create account v2 (#3281)
## What does this change?
This pr adds the ability to register as well as sign in when using the
`profile.theguardian.com/signin` url. This should hopefully help out
where users that thought they had an account but didn't try to sign in
via the /signin page. Previously these users would not have been sent an
email which can be confusing.