Maintain query string args in returnUrl for observer subdomain (#3290)
We are implementing serving the checkout for Observer subs on
observer.theguardian.com (with different branding). If you're signed in
and click the link on the checkout to sign out, the flow is currently
broken because the `returnUrl` passed to gateway's `signOut` endpoint
has its query string args stripped out. So the URL you land back on no
longer contains the product/ratePlan and shows an error.

This change ensures we keep the query string args.