Add exception for '/welcome/complete-account' (#3291)
<!-- See
https://github.com/guardian/recommendations/blob/main/pull-requests.md
for recommendations on raising and reviewing pull requests. -->

## What does this change?
This PR fixes a routing isssue where jobs users were unable to complete
their account after registration. The fix adds
`/welcome/complete-account` as an exception to the fromURI redirect
logic, ensuring users are always routed to the account completion page
when needed, regardless of the fromURI parameter.

<!-- A PR should have enough detail to be understandable far in the
future. e.g what is the problem/why is the change needed, how does it
solve it and any questions or points of discussion. Prefer copying
information from a Trello card over linking to it; the card may not
always exist and reviewers may not have access to the board. -->

## How to test
This has been tested in CODE by using the Jobs CODE environment to sign
in.
<!-- Provide instructions to help others verify the change. This could
take the form of "On PROD, do X and witness Y. On this branch, do X and
witness Z. " -->